move accf fixtures into test module, importorskip entire module
